requirements.txt
absl-py==2.1.0
agentops==0.3.21
aiocache==0.12.3
aiofiles==23.2.1
aiohttp==3.8.6
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anthropic==0.18.1
anyio==4.7.0
anytree==2.12.1
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
asyncer==0.0.7
attrs==24.3.0
azure-core==1.32.0
azure-identity==1.19.0
backoff==2.2.1
bce-python-sdk==0.9.25
bcrypt==4.2.1
beautifulsoup4==4.12.3
bidict==0.23.1
boto3==1.34.162
botocore==1.34.162
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.0
camel-converter==4.0.1
certifi==2024.12.14
cffi==1.17.1
chainlit==2.2.1
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.4.1
chevron==0.14.0
chinadata==0.1.3
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.8
cloudpickle==3.1.0
cohere==5.13.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.0
cryptography==43.0.3
curl_cffi==0.7.4
cycler==0.12.1
dashscope==1.19.3
dataclasses-json==0.6.7
debugpy==1.8.11
decorator==5.1.1
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.9
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
Django==4.2.17
dnspython==2.7.0
docx2txt==0.8
duckduckgo_search==7.2.1
durationpy==0.9
elastic-transport==8.17.0
elasticsearch==8.17.0
environs==9.5.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.1.0
faiss-cpu==1.7.4
Farama-Notifications==0.0.4
fastapi==0.115.6
fastavro==1.10.0
fastjsonschema==2.21.1
fasttext==0.9.3
filelock==3.16.1
filetype==1.2.0
fire==0.4.0
flatbuffers==24.12.23
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.12.0
future==1.0.0
fuzzywuzzy==0.18.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.12
gitignore_parser==0.1.9
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.24.0
google-auth==2.37.0
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
gritql==0.1.5
grpcio==1.67.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
gymnasium==0.29.1
h11==0.14.0
h2==4.1.0
h5py==3.13.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
imap-tools==1.5.0
importlib_metadata==8.4.0
importlib_resources==6.5.2
ipykernel==6.27.1
ipython==8.17.2
ipywidgets==8.1.1
isodate==0.7.2
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-spec==0.2.4
jsonschema-specifications==2023.7.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
keras==3.8.0
kiwisolver==1.4.7
kubernetes==31.0.0
lancedb==0.4.0
langchain-cli==0.0.35
langchain-community==0.3.14
langchain-core==0.3.30
langchain-text-splitters==0.3.5
langserve==0.3.1
langsmith==0.2.11
Lazify==0.4.0
lazy-object-proxy==1.10.0
Levenshtein==0.26.1
libclang==18.1.1
libcst==1.0.1
literalai==0.1.103
llama-index-core==0.10.15
llama-index-embeddings-azure-openai==0.1.6
llama-index-embeddings-gemini==0.1.6
llama-index-embeddings-ollama==0.1.2
llama-index-embeddings-openai==0.1.5
llama-index-embeddings-zhipuai==0.2.0
llama-index-llms-azure-openai==0.1.4
llama-index-llms-openai==0.1.8
llama-index-postprocessor-cohere-rerank==0.1.4
llama-index-postprocessor-colbert-rerank==0.1.1
llama-index-postprocessor-flag-embedding-reranker==0.1.2
llama-index-readers-file==0.1.4
llama-index-retrievers-bm25==0.1.3
llama-index-vector-stores-chroma==0.1.6
llama-index-vector-stores-elasticsearch==0.1.6
llama-index-vector-stores-faiss==0.1.1
llama-index-vector-stores-milvus==0.1.23
llamaindex-py-client==0.1.19
loguru==0.6.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
meilisearch==0.21.0
milvus-lite==2.4.11
ml-dtypes==0.4.1
mmh3==5.0.1
monotonic==1.6
more-itertools==10.5.0
motor==3.6.0
mpmath==1.3.0
msal==1.31.1
msal-extensions==1.2.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.9.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.39.0
openapi-core==0.18.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
opt_einsum==3.4.0
optree==0.14.0
orjson==3.10.14
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.1.1
parameterized==0.9.0
parse==1.20.2
parso==0.8.4
pathable==0.4.3
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
playwright==1.49.1
portalocker==2.10.1
posthog==3.7.5
prance==23.6.21.0
primp==0.10.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.8
psutil==6.0.0
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
py==1.11.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pyee==12.0.0
pygame==2.6.1
Pygments==2.18.0
PyJWT==2.8.0
pylance==0.9.0
pymilvus==2.4.9
pymongo==4.9.2
PyMuPDF==1.25.2
pyparsing==3.2.1
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.11.2
python-Levenshtein==0.26.1
python-multipart==0.0.18
python-socketio==5.12.1
pytz==2024.2
PyYAML==6.0.1
pyzmq==26.2.0
qdrant-client==1.7.0
qianfan==0.4.12.2
rank-bm25==0.2.2
RapidFuzz==3.12.1
ratelimiter==1.2.0.post0
redis==5.0.8
referencing==0.30.2
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rfc3339-validator==0.1.4
rich==13.6.0
rpds-py==0.22.3
rsa==4.9
ruamel.yaml==0.18.8
ruamel.yaml.clib==0.2.12
s3transfer==0.10.4
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.13.1
seaborn==0.13.2
selenium==4.27.1
semantic-kernel==0.4.3.dev0
semver==3.0.2
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
spark-ai-python==0.3.31
SQLAlchemy==2.0.37
sqlparse==0.5.3
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.41.3
striprtf==0.0.26
sympy==1.13.1
syncer==2.0.3
ta==0.10.2
tenacity==8.2.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.2
tqdm==4.66.2
traitlets==5.14.3
transformers==4.48.0
tree-sitter==0.23.2
tree-sitter-python==0.23.6
trio==0.28.0
trio-websocket==0.11.1
tushare==1.4.18
typer==0.9.0
types-requests==2.31.0.6
types-urllib3==1.26.25.14
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2024.2
ujson==5.10.0
uptrace==1.29.0
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
volcengine-python-sdk==1.0.119
watchfiles==0.20.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.15.0
wsproto==1.2.0
xgboost==2.1.4
yarl==1.18.3
zhipuai==2.1.5.20241204
zipp==3.21.0
